version bump

+++ tools/generate_project_mac.sh
@@ -5,8 +5,14 @@ cd $(dirname "$0")
 
 cd ..
 root_path=$(pwd)
-rm -rf build
-mkdir build
+
+if [ ! -d "build/" ]; then
+	echo "INFO: First time setup will take longer as the dependencies need to be downloaded and compiled."
+else
+	rm -rf build
+fi
+
+mkdir "build"
 cd build
 
 build_type=""
